refatoração da rota de login e ajueste nos erros

+++ src/use-cases/delete-user.ts
@@ -1,3 +1,4 @@
+import { UserNotExist } from "@/erros/cheking-cases/cheking-user-exist"
 import { UserRepository } from "@/repositories/users-repository"
 
 interface deleteUserRequest {
@@ -8,6 +9,12 @@ export class DeleteUser {
     constructor(private userRepository: UserRepository) { }
 
     async execute({ id }: deleteUserRequest) {
+
+        const userExist = this.userRepository.exists(id)
+        if (!userExist) {
+            throw new UserNotExist()
+        }
+
         await this.userRepository.delete(
             id
         )
